+++ CHANGELOG.md
@@ -6,6 +6,7 @@ v2.0.x
 
 - Breaking changes to migrate dictionaries to ```NameValueCollection```
 - Retarget to include .NET Framework 4.8 and .NET 7.0
+- Reintroduce ```HttpRequest``` methods for checking existence of and retrieving query or header values
 
 ## Previous Versions